remove calculateAbsorption and move it's logic into calculateDamage for simplicity
